Reorganize THD directory structure
