+++ torch/lib/THD/THD.h
@@ -6,8 +6,8 @@
 #define THD_API
 #endif
 
-#include "master/Master.h"
-#include "master/State.h"
-#include "master/THDTensor.h"
+#include "master_worker/master/Master.h"
+#include "master_worker/master/State.h"
+#include "master_worker/master/THDTensor.h"
 
-#include "worker/Worker.h"
+#include "master_worker/worker/Worker.h"
