--- torch/lib/THD/_THD.h
@@ -1,19 +1,0 @@
-#pragma once
-
-#import "common/_Channel.h"
-#include "common/_CommandChannel.h"
-#include "common/_DataChannel.h"
-#ifdef WITH_MPI
-#include "common/_DataChannelMPI.h"
-#endif
-#include "common/_DataChannelTCP.h"
-#include "common/_Functions.h"
-#include "common/_RPC.h"
-#include "common/_THTensor.h"
-#include "common/_Tensor.h"
-
-#include "master/_State.h"
-
-#include "worker/_Dispatch.h"
-
-#include "THD.h"
