+++ torch/lib/THD/base/channels/DataChannelTCP.cpp
@@ -1,4 +1,4 @@
-#include "../_THD.h"
+#include "DataChannelTCP.hpp"
 
 #include <arpa/inet.h>
 #include <sys/poll.h>
