+++ torch/lib/THD/base/tensors/THTensor.cpp
@@ -1,4 +1,4 @@
-#include "../_THD.h"
+#include "THTensor.hpp"
 
 #include "generic/THTensor.cpp"
 #include <TH/THGenerateAllTypes.h>
