--- torch/lib/THD/master/THDTensor.cpp
@@ -1,1 +1,0 @@
-#include "../_THD.h"
