+++ torch/lib/THD/master_worker/common/ByteArray.cpp
@@ -1,4 +1,4 @@
-#include "_ByteArray.h"
+#include "ByteArray.hpp"
 
 #include <cstdlib>
 #include <cstddef>
