+++ torch/lib/THD/master_worker/common/ByteArray.hpp
