+++ torch/lib/THD/master_worker/common/CommandChannel.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "_RPC.h"
+#include "RPC.hpp"
 
 #include <memory>
 #include <string>
