+++ torch/lib/THD/master_worker/common/Functions.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <stdint.h>
+#include <cstdint>
 
 namespace thd {
