+++ torch/lib/THD/master_worker/common/RPC-inl.hpp
@@ -1,4 +1,3 @@
-#include "_ByteArray.h"
 
 namespace thd { namespace rpc { namespace detail {
 ////////////////////////////////////////////////////////////////////////////////
