+++ torch/lib/THD/master_worker/master/Master.h
