+++ torch/lib/THD/master_worker/master/State.hpp
