+++ torch/lib/THD/master_worker/worker/Dispatch.cpp
@@ -1,7 +1,7 @@
 #include <unordered_map>
 #include <string>
 
-#include "../_THD.h"
+#include "../common/Functions.hpp"
 
 namespace thd {
 namespace worker {
