+++ torch/lib/THD/master_worker/worker/Dispatch.hpp
