+++ torch/lib/THD/master_worker/worker/Worker.h
@@ -1,3 +1,5 @@
 #pragma once
 
+#include "THD.h"
+
 THD_API void THDWorkerMain();
