+++ torch/lib/THD/test/data_channel_tcp_accept_timeout.cpp
@@ -1,4 +1,4 @@
-#include "../_THD.h"
+#include "../base/channels/DataChannelTCP.hpp"
 
 #include <cassert>
 #include <iostream>
