+++ torch/lib/THD/test/data_channel_tcp_slow_master.cpp
@@ -1,4 +1,5 @@
-#include "../_THD.h"
+#include "../base/channels/DataChannelTCP.hpp"
+#include "../base/tensors/THTensor.hpp"
 
 #include <cassert>
 #include <iostream>
